week2/day6/event_management_system/requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
python-dotenv
flask-migrate
flask-wtf
WTForms
